% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_normality_function.r
\name{assess_normality}
\alias{assess_normality}
\title{Create a customized histogram figure}
\usage{
assess_normality(x, ...)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired}

\item{...}{one of:
a vector giving the breakpoints between histogram cells,
a function to compute the vector of breakpoints,
a single number giving the number of cells for the histogram,
a character string naming an algorithm to compute the number of cells,
a function to compute the number of cells.}
}
\value{
a histogram figure
}
\description{
Create a customized histogram figure
}
\examples{
data(faithful)
assess_normality(x = faithful$eruptions, main = "Old Faithful data")
}
