% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_journaltable_function.r
\name{create.journaltable}
\alias{create.journaltable}
\title{Create a customized summary table with journal article aesthetics}
\usage{
create.journaltable(data, y, x, b)
}
\arguments{
\item{data}{A data frame}

\item{y}{A categorical variable name (quoted or unquoted) in data. Summary statistics will
be calculated separately for each level of the variable.}

\item{x}{continuous variable/s}

\item{b}{categorical variable/s}
}
\value{
a summary table with journal article aesthetic for all variables in dataset
}
\description{
Create a customized summary table with journal article aesthetics
}
\examples{
create.journaltable(data=trial,y='trt', x='age', b='death')
}
