% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_journaltable_function.r
\name{create_journaltable}
\alias{create_journaltable}
\title{Create a customized summary table}
\usage{
create_journaltable(data, y, b, x)
}
\arguments{
\item{data}{A data frame}

\item{y}{A categorical variable name (quoted or unquoted) in data.
Summary statistics will be calculated separately for each level of the variable.}

\item{b}{categorical variable/s}

\item{x}{continuous variable/s}
}
\value{
a customized summary table for all variables in data set with journal
article aesthetics
}
\description{
Create a customized summary table
}
\examples{
\dontrun{
create_journaltable(data = gtsummary::trial, y = "trt", b = ("death"), x = ("age"))
}

}
