% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_function.r
\name{download_data}
\alias{download_data}
\title{Read in public data set URL file from github's website}
\usage{
download_data(urlfile)
}
\arguments{
\item{urlfile}{github repository link where the CSV file of dataset is located}
}
\value{
output list of variable names, type, and observations
}
\description{
Read in public data set URL file from github's website
}
\examples{
download_data(urlfile ="https://raw.githubusercontent.com/RotanaRad/rotsurpackage/main/GLOW_new.csv")

}
