% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_selection_function.r
\name{stepwise_selection}
\alias{stepwise_selection}
\title{Select predicotor that create best fit model using forward selection methods}
\usage{
stepwise_selection(data, formula)
}
\arguments{
\item{data}{a data frame}

\item{formula}{specify dependent and dependent variables for glm formula}
}
\value{
a  list of candidiate variables, list of added variables, and stepwise
summary table
}
\description{
Select predicotor that create best fit model using forward selection methods
}
\examples{
stepwise_selection(trial, response~.)
}
