% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_finalmodel_function.r
\name{summary_finalmodel}
\alias{summary_finalmodel}
\title{Create a summary table with results of a glm model}
\usage{
summary_finalmodel(data, formula)
}
\arguments{
\item{data}{a data frame}

\item{formula}{specify dependent and independent variables for glm formula}
}
\value{
summaries of the results of various model fitting functions.
}
\description{
Create a summary table with results of a glm model
}
\examples{
summary_finalmodel(data = gtsummary::trial, (substitute(response ~ age + factor(death))))
}
