% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_journaltable_function.r
\name{create_journaltable}
\alias{create_journaltable}
\title{Create a customized summary table}
\usage{
create_journaltable(data, y)
}
\arguments{
\item{data}{A data frame}

\item{y}{}
}
\description{
Create a customized summary table
}
\examples{
create_journaltable(data =infert, y = 'case')
}
